rtp: link uri handing if not in core
A patch has been submitted to GStreamer core to allow picking up
parameters of an object that are in the query part of an URI. This is
not yet included in GStreamer Core.

As long as it is not in, the relevant code will be linked in
conditionally.
See: https://bugzilla.gnome.org/show_bug.cgi?id=779765.